💄 Allow meta on new tags ui
