+++ static/app/components/events/eventTags/eventTagsTree.tsx
@@ -1,21 +1,13 @@
-import {Fragment, useMemo, useRef, useState} from 'react';
+import {Fragment, useMemo, useRef} from 'react';
 import styled from '@emotion/styled';
-import * as qs from 'query-string';
 
-import {openNavigateToExternalLinkModal} from 'sentry/actionCreators/modal';
-import {navigateTo} from 'sentry/actionCreators/navigation';
-import {DropdownMenu} from 'sentry/components/dropdownMenu';
-import Version from 'sentry/components/version';
-import VersionHoverCard from 'sentry/components/versionHoverCard';
-import {IconEllipsis} from 'sentry/icons';
-import {t} from 'sentry/locale';
+import EventTagsTreeRow, {
+  type EventTagsTreeRowProps,
+} from 'sentry/components/events/eventTags/eventTagsTreeRow';
 import {space} from 'sentry/styles/space';
 import type {EventTag} from 'sentry/types';
 import type {Event} from 'sentry/types/event';
-import {generateQueryWithTag, isUrl} from 'sentry/utils';
 import {useDimensions} from 'sentry/utils/useDimensions';
-import useOrganization from 'sentry/utils/useOrganization';
-import useRouter from 'sentry/utils/useRouter';
 
 const MAX_TREE_DEPTH = 4;
 const INVALID_BRANCH_REGEX = /\.{2,}/;
@@ -36,15 +28,6 @@ interface TagTreeColumnData {
   columns: React.ReactNode[];
   runningTotal: number;
   startIndex: number;
-}
-
-interface TagTreeRowProps {
-  content: TagTreeContent;
-  event: Event;
-  projectSlug: string;
-  tagKey: string;
-  isLast?: boolean;
-  spacerCount?: number;
 }
 
 interface EventTagsTreeProps {
@@ -89,164 +72,17 @@ function addToTagTree(
   return tree;
 }
 
-function TagTreeRow({
-  event,
-  content,
-  tagKey,
-  spacerCount = 0,
-  isLast = false,
-  projectSlug,
-}: TagTreeRowProps) {
-  const organization = useOrganization();
-  const router = useRouter();
-  const [isVisible, setIsVisible] = useState(false);
-  const originalTag = content.originalTag;
-
-  if (!originalTag) {
-    return (
-      <TreeRow data-test-id="tag-tree-row">
-        <TreeKeyTrunk spacerCount={spacerCount}>
-          {spacerCount > 0 && (
-            <Fragment>
-              <TreeSpacer spacerCount={spacerCount} isLast={isLast} />
-              <TreeBranchIcon />
-            </Fragment>
-          )}
-          <TreeKey>{tagKey}</TreeKey>
-        </TreeKeyTrunk>
-        <TreeValueTrunk />
-      </TreeRow>
-    );
-  }
-
-  const referrer = 'event-tags-tree';
-  const query = generateQueryWithTag({referrer}, originalTag);
-  const searchQuery = `?${qs.stringify(query)}`;
-
-  return (
-    <TreeRow data-test-id="tag-tree-row">
-      <TreeKeyTrunk spacerCount={spacerCount}>
-        {spacerCount > 0 && (
-          <Fragment>
-            <TreeSpacer spacerCount={spacerCount} isLast={isLast} />
-            <TreeBranchIcon />
-          </Fragment>
-        )}
-        <TreeSearchKey aria-hidden>{originalTag.key}</TreeSearchKey>
-        <TreeKey title={originalTag.key}>{tagKey}</TreeKey>
-      </TreeKeyTrunk>
-      <TreeValueTrunk>
-        <TreeValue>
-          {originalTag.key === 'release' ? (
-            <VersionHoverCard
-              organization={organization}
-              projectSlug={projectSlug}
-              releaseVersion={content.value}
-              showUnderline
-              underlineColor="linkUnderline"
-            >
-              <Version version={content.value} truncate />
-            </VersionHoverCard>
-          ) : (
-            content.value
-          )}
-        </TreeValue>
-        <TreeValueDropdown
-          preventOverflowOptions={{padding: 4}}
-          className={isVisible ? '' : 'invisible'}
-          position="bottom-end"
-          size="xs"
-          onOpenChange={isOpen => setIsVisible(isOpen)}
-          triggerProps={{
-            'aria-label': t('Tag Actions Menu'),
-            icon: <IconEllipsis />,
-            showChevron: false,
-            className: 'tag-button',
-          }}
-          items={[
-            {
-              key: 'view-events',
-              label: t('View other events with this tag value'),
-              hidden: !event.groupID,
-              onAction: () => {
-                navigateTo(
-                  `/organizations/${organization.slug}/issues/${event.groupID}/events/${searchQuery}`,
-                  router
-                );
-              },
-            },
-            {
-              key: 'view-issues',
-              label: t('View issues with this tag value'),
-              onAction: () => {
-                navigateTo(
-                  `/organizations/${organization.slug}/issues/${searchQuery}`,
-                  router
-                );
-              },
-            },
-            {
-              key: 'release',
-              label: t('View this release'),
-              hidden: originalTag.key !== 'release',
-              onAction: () => {
-                navigateTo(
-                  `/organizations/${organization.slug}/releases/${encodeURIComponent(
-                    content.value
-                  )}/`,
-                  router
-                );
-              },
-            },
-            {
-              key: 'transaction',
-              label: t('View this transaction'),
-              hidden: originalTag.key !== 'transaction',
-              onAction: () => {
-                const transactionQuery = qs.stringify({
-                  project: event.projectID,
-                  transaction: content.value,
-                  referrer,
-                });
-                navigateTo(
-                  `/organizations/${organization.slug}/performance/summary/?${transactionQuery}`,
-                  router
-                );
-              },
-            },
-            {
-              key: 'replay',
-              label: t('View this replay'),
-              hidden: originalTag.key !== 'replay_id' && originalTag.key !== 'replayId',
-              onAction: () => {
-                const replayQuery = qs.stringify({referrer});
-                navigateTo(
-                  `/organizations/${organization.slug}/replays/${encodeURIComponent(content.value)}/?${replayQuery}`,
-                  router
-                );
-              },
-            },
-            {
-              key: 'external-link',
-              label: t('Visit this external link'),
-              hidden: !isUrl(content.value),
-              onAction: () => {
-                openNavigateToExternalLinkModal({linkText: content.value});
-              },
-            },
-          ]}
-        />
-      </TreeValueTrunk>
-    </TreeRow>
-  );
-}
-
 /**
  * Function to recursively create a flat list of all rows to be rendered for a given TagTree
  * @param props The props for rendering the root of the TagTree
  * @returns A list of TagTreeRow components to be rendered in this tree
  */
-function getTagTreeRows({tagKey, content, spacerCount = 0, ...props}: TagTreeRowProps) {
+function getTagTreeRows({
+  tagKey,
+  content,
+  spacerCount = 0,
+  ...props
+}: EventTagsTreeRowProps) {
   const subtreeTags = Object.keys(content.subtree);
   const subtreeRows = subtreeTags.reduce((rows, tag, i) => {
     const branchRows = getTagTreeRows({
@@ -260,7 +96,7 @@ function getTagTreeRows({tagKey, content, spacerCount = 0, ...props}: TagTreeRow
   }, []);
 
   return [
-    <TagTreeRow
+    <EventTagsTreeRow
       key={`${tagKey}-${spacerCount}`}
       tagKey={tagKey}
       content={content}
@@ -367,96 +203,6 @@ const TreeColumn = styled('div')`
   &:not(:last-child) {
     border-right: 1px solid ${p => p.theme.innerBorder};
     padding-right: ${space(2)};
-  }
-`;
-
-const TreeRow = styled('div')`
-  border-radius: ${space(0.5)};
-  padding-left: ${space(1)};
-  position: relative;
-  display: grid;
-  align-items: center;
-  grid-column: span 2;
-  grid-template-columns: subgrid;
-  :nth-child(odd) {
-    background-color: ${p => p.theme.backgroundSecondary};
-  }
-  .invisible {
-    visibility: hidden;
-  }
-  &:hover,
-  &:active {
-    .invisible {
-      visibility: visible;
-    }
-  }
-`;
-
-const TreeSpacer = styled('div')<{isLast: boolean; spacerCount: number}>`
-  grid-column: span 1;
-  /* Allows TreeBranchIcons to appear connected vertically */
-  border-right: 1px solid ${p => (!p.isLast ? p.theme.border : 'transparent')};
-  margin-right: -1px;
-  height: 100%;
-`;
-
-const TreeBranchIcon = styled('div')`
-  border: 1px solid ${p => p.theme.border};
-  border-width: 0 0 1px 1px;
-  border-radius: 0 0 0 5px;
-  grid-column: span 1;
-  height: 12px;
-  align-self: start;
-  margin-right: ${space(0.5)};
-`;
-
-const TreeKeyTrunk = styled('div')<{spacerCount: number}>`
-  grid-column: 1 / 2;
-  display: grid;
-  height: 100%;
-  align-items: center;
-  grid-template-columns: ${p =>
-    p.spacerCount > 0 ? `${(p.spacerCount - 1) * 20 + 3}px 1rem 1fr` : '1fr'};
-`;
-
-const TreeValueTrunk = styled('div')`
-  grid-column: 2 / 3;
-  display: grid;
-  height: 100%;
-  align-items: center;
-  min-height: 22px;
-  grid-template-columns: 1fr auto;
-  grid-column-gap: ${space(0.5)};
-`;
-
-const TreeValue = styled('div')`
-  font-family: ${p => p.theme.text.familyMono};
-  font-size: ${p => p.theme.fontSizeSmall};
-  word-break: break-word;
-  grid-column: span 1;
-`;
-
-const TreeKey = styled(TreeValue)`
-  color: ${p => p.theme.gray300};
-`;
-
-/**
- * Hidden element to allow browser searching for exact key name
- */
-const TreeSearchKey = styled('span')`
-  font-size: 0;
-  position: absolute;
-`;
-
-const TreeValueDropdown = styled(DropdownMenu)`
-  margin: 1px;
-  height: 20px;
-  .tag-button {
-    height: 20px;
-    min-height: 20px;
-    padding: ${space(0)} ${space(0.75)};
-    border-radius: ${space(0.5)};
-    z-index: 0;
   }
 `;
